requirements.txt
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
python-dotenv>=1.0.0
Pillow>=10.0.0